Refactor mouse centering logic in Controller class
Refactored the `Controller::update` method to call a new `centerMouse` method instead of directly setting the mouse position to the center of the screen. Added the `centerMouse` method to `Controller.cpp`.
Updated `State_Game::activate` in `State_Game.cpp` to call `centerMouse` when the game state is activated. Modified several binary files due to recompilation.